Different fix for the wrong screen problem.
